requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\github\Memehub\controllers\controller\app.py: 4
Flask == 1.1.2

# C:\github\Memehub\controllers\controller\generated\models.py: 4
# C:\github\Memehub\controllers\controller\reddit\functions\database.py: 5
# C:\github\Memehub\controllers\controller\reddit\scorer.py: 4,5
# C:\github\Memehub\controllers\controller\redis\reddit.py: 14
# C:\github\Memehub\controllers\controller\utils\build_features.py: 5
SQLAlchemy == 1.3.20

# C:\github\Memehub\controllers\controller\constants.py: 1
# C:\github\Memehub\controllers\controller\reddit\controller.py: 1
arrow == 0.17.0

# C:\github\Memehub\controllers\scripts\hive.py: 3,4,5,6,7,8,9
beem == 0.24.20

# C:\github\Memehub\controllers\controller\reddit\controller.py: 12
# C:\github\Memehub\controllers\controller\reddit\functions\praw_mp.py: 4
billiard == 3.6.3.0

# C:\github\Memehub\controllers\config\flask.py: 3
# C:\github\Memehub\controllers\controller\app.py: 1
celery == 5.0.4

# C:\github\Memehub\controllers\cli\cli.py: 1
# C:\github\Memehub\controllers\cli\commands\cmd_gen.py: 3
# C:\github\Memehub\controllers\cli\commands\cmd_load_redisai.py: 1
# C:\github\Memehub\controllers\cli\commands\cmd_scripts.py: 1
click == 7.1.2

# C:\github\Memehub\controllers\config\flask.py: 4
# C:\github\Memehub\controllers\controller\app.py: 3
# C:\github\Memehub\controllers\controller\reddit\functions\praw_mp.py: 6
# C:\github\Memehub\controllers\scripts\hive.py: 10
python-decouple == 3.3

# C:\github\Memehub\controllers\controller\generated\models.py: 3
flask-sqlalchemy == 2.4.4

# C:\github\Memehub\controllers\config\flask.py: 5
get-docker-secret == 1.0.1

# C:\github\Memehub\controllers\controller\reddit\controller.py: 4,13
# C:\github\Memehub\controllers\controller\reddit\functions\dataframe.py: 5
# C:\github\Memehub\controllers\controller\reddit\functions\misc.py: 2
# C:\github\Memehub\controllers\controller\utils\build_features.py: 1
# C:\github\Memehub\controllers\controller\utils\extract_image.py: 3
# C:\github\Memehub\controllers\controller\utils\image_text.py: 2
# C:\github\Memehub\controllers\controller\utils\save_image.py: 4
numpy == 1.19.2

# C:\github\Memehub\controllers\controller\reddit\controller.py: 2
# C:\github\Memehub\controllers\controller\reddit\functions\misc.py: 1
# C:\github\Memehub\controllers\controller\utils\extract_image.py: 4
# C:\github\Memehub\controllers\controller\utils\image_text.py: 1
# C:\github\Memehub\controllers\controller\utils\save_image.py: 3
opencv-python-headless == 4.4.0.46

# C:\github\Memehub\controllers\controller\reddit\controller.py: 6
# C:\github\Memehub\controllers\controller\reddit\functions\dataframe.py: 4
# C:\github\Memehub\controllers\controller\reddit\scorer.py: 3
pandas == 1.1.3

# C:\github\Memehub\controllers\controller\reddit\functions\praw_mp.py: 7,8
praw == 7.1.0

# C:\github\Memehub\controllers\controller\reddit\controller.py: 7
# C:\github\Memehub\controllers\controller\utils\image_text.py: 3
pytesseract == 0.3.6

# C:\github\Memehub\controllers\controller\reddit\controller.py: 8
# C:\github\Memehub\controllers\controller\redis\reddit.py: 16
redis == 3.5.3

# C:\github\Memehub\controllers\controller\rai.py: 3
# C:\github\Memehub\controllers\controller\reddit\controller.py: 9
# C:\github\Memehub\controllers\controller\utils\build_features.py: 2
redisai == 1.0.1

# C:\github\Memehub\controllers\controller\reddit\controller.py: 10
# C:\github\Memehub\controllers\controller\reddit\functions\pushshift.py: 4
# C:\github\Memehub\controllers\controller\utils\extract_image.py: 2
# C:\github\Memehub\controllers\controller\utils\image_text.py: 4
# C:\github\Memehub\controllers\controller\utils\save_image.py: 5
requests == 2.24.0

# C:\github\Memehub\controllers\setup.py: 1
setuptools == 50.3.1

# C:\github\Memehub\controllers\controller\reddit\controller.py: 14
# C:\github\Memehub\controllers\controller\utils\build_features.py: 6
tqdm == 4.50.2

celery-singleton == 0.3.0

ml2rt == 0.2.0
boto3

torch
torchvision
matplotlib

beautifulsoup4 == 4.9.3
alembic == 1.4.3

# notebook=6.1.4
# jupyterlab
tornado==6.1
# not found by pigar
retry
sqlacodegen
flask-sqlacodegen
gunicorn == 20.0.4
psycopg2-binary == 2.8.4
flask-cors
# flower == 0.9.3
lxml == 4.5.0
Pillow == 7.1.2
watchdog